# iteh code/eventapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventHub API",
        "description": "API dokumentacija za EventHub aplikaciju",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Prikaz svih notifikacija ulogovanog korisnika",
                "operationId": "afdfd7d2bbf2fcdaf5b8c1e1e07e4eec",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena lista notifikacija"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Kreiranje nove notifikacije",
                "operationId": "c1bc4a871653ba54a228c6fbf695580a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idEvent",
                                    "message",
                                    "type"
                                ],
                                "properties": {
                                    "idEvent": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Podsetnik: događaj počinje sutra u 18h"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "REMINDER"
                                    },
                                    "seen": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notifikacija uspešno kreirana"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Prikaz jedne notifikacije",
                "operationId": "80f8f96dbc2dcd82aae78323704e90bc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notifikacije (idNotification)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno vraćena notifikacija"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Notifikacija nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Izmena postojeće notifikacije",
                "operationId": "dd3c70b57d876ff46a96ec571f60962c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notifikacije (idNotification)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idEvent": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Ažurirana poruka notifikacije"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "UPDATE"
                                    },
                                    "seen": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notifikacija uspešno izmenjena"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Notifikacija nije pronađena"
                    },
                    "422": {
                        "description": "Greška validacije"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Brisanje notifikacije",
                "operationId": "731ec9a91c011379af0c8c0af66bcfe4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID notifikacije (idNotification)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifikacija uspešno obrisana"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "404": {
                        "description": "Notifikacija nije pronađena"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}